addressing suggestions; moving PropertyView component to a new file
